Create Marketplace_BLOCKS_Data_V1.sol
This updated contract includes the JobRequirement struct and an additional mapping called jobRequirements to store job requirements data. The createJob function now also accepts job requirement data as parameters and stores them in the jobRequirements mapping.

Now you can store job requirements on-chain with the rest of the job data.